+++ cli/lib/src/task_runner/observatory_discovery.dart
@@ -28,16 +28,26 @@ class ObservatoryDiscovery with Log {
     await _deviceLogSubscription.cancel();
   }
 
+  /// Matches strings like:
+  /// 1. Observatory listening on http://127.0.0.1:8181/bKrWeCYueRo=/
+  /// 2. A Dart VM service is listening on http://127.0.0.1:8181/abc123=/
+  /// 
+  /// This function used to only match string 1 above, but as of 
+  /// Flutter 2.13.0-0.1.pre (April 2022), the Dart team changed it to string 2. 
+  /// (https://github.com/dart-lang/sdk/issues/46756)
+  /// 
+  /// This function will match either string. Keep in mind that moving forward 
+  /// only string 2 will need to be supported.
   Match? _getPatternMatch(String line) {
-    final r = RegExp(RegExp.escape(serviceName) +
-        r' listening on ((http|//)[a-zA-Z0-9:/=_\-\.\[\]]+)');
+    final r = RegExp(
+        r'(Observatory|The Dart VM service is) listening on ((http|//)[a-zA-Z0-9:/=_\-\.\[\]]+)');
     return r.firstMatch(line);
   }
 
   Uri? _getObservatoryUri(String line) {
     final match = _getPatternMatch(line);
     if (match != null) {
-      return Uri.parse(match[1]!);
+      return Uri.parse(match[2]!);
     }
     return null;
   }