+++ packages/monarch/lib/src/monarch_app/standard_themes.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'channel_methods.dart';
+import 'channel_argument.dart';
 import 'monarch_data.dart';
 
 final standardMetaThemes = [